+++ src/sysdolphin/baselib/objalloc.c
@@ -12,39 +12,39 @@ static HSD_ObjAllocData* alloc_datas;
 
 void HSD_ObjSetHeap(u32 size, void* ptr)
 {
-    obj_heap.curr = (u32) ptr;
-    obj_heap.top = (u32) ptr;
+    obj_heap.curr = (uintptr_t) ptr;
+    obj_heap.top = (uintptr_t) ptr;
     obj_heap.remain = size;
     obj_heap.size = size;
 }
 
 s32 HSD_ObjAllocAddFree(HSD_ObjAllocData* data, u32 num)
 {
-    u32 computed_start;
-    u32 pool_end;
-    u32 pool_size;
+    uintptr_t computed_start;
+    uintptr_t pool_end;
+    uintptr_t pool_size;
     u8* pool_start;
 
     u8 _[4];
 
     HSD_ASSERT(0xEE, data);
-    pool_size = data->size * num;
+    pool_size = (uintptr_t) data->size * num;
     if (obj_heap.top != 0) {
         pool_end = obj_heap.top + obj_heap.size;
-        computed_start = (obj_heap.curr + data->align) & ~data->align;
+        computed_start = (obj_heap.curr + data->align) & ~((uintptr_t) data->align);
         pool_start = (void*) computed_start;
         if (computed_start > pool_end) {
             return 0;
         }
-        if (pool_end - (u32) pool_start < pool_size) {
-            pool_size = pool_end - (u32) pool_start -
-                        (pool_end - (u32) pool_start) % data->size;
+        if (pool_end - (uintptr_t) pool_start < pool_size) {
+            pool_size = pool_end - (uintptr_t) pool_start -
+                        (pool_end - (uintptr_t) pool_start) % data->size;
         }
         num = pool_size / data->size;
         if (num == 0) {
             return 0;
         }
-        obj_heap.curr = (u32) pool_start + pool_size;
+        obj_heap.curr = (uintptr_t) pool_start + pool_size;
         obj_heap.remain = pool_end - obj_heap.curr;
     } else {
         pool_start = HSD_MemAlloc(pool_size);
